AgriPredict/requirements.txt
# requirements-windows.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.30.0
plotly>=5.15.0
shap>=0.43.0
joblib>=1.3.0
pillow>=9.0.0